+++ src/sequentially.ts
@@ -1,6 +1,6 @@
 import { array } from './array';
 import { isPromise } from './is_promise';
-import { NFn, NArray } from './utilities';
+import { NFn, NArray, If } from './utilities';
 
 const concat = <V, A extends ReadonlyArray<V>>(
   aggregator: A,
@@ -9,6 +9,10 @@ const concat = <V, A extends ReadonlyArray<V>>(
   isPromise<V>(result)
     ? result.then((value) => aggregator.concat(value))
     : aggregator.concat(result);
+
+type SequentiallyParameters<V extends ReadonlyArray<any>> = NArray.Flatten<
+  NArray.TrimLastEmpty<NFn.ParametersOf<V>>
+>;
 
 /**
  * Independently invokes functions with their parameters
@@ -20,14 +24,26 @@ export const sequentially =
     ...fns: V
   ) =>
   (
-    ...values: NArray.Flatten<NArray.TrimLastEmpty<NFn.ParametersOf<V>>>
+    ...values: If<
+      NArray.IsSameInnerType<SequentiallyParameters<V>>,
+      SequentiallyParameters<V> | [NArray.First<SequentiallyParameters<V>>],
+      SequentiallyParameters<V>
+    >
   ): NFn.IsAsyncIn<V> extends true
     ? Promise<NFn.ReturnTypesOf<V>>
-    : NFn.ReturnTypesOf<V> =>
-    fns.reduce((waiter, fn, index) => {
-      const parameters = array((values as any[])[index]);
+    : NFn.ReturnTypesOf<V> => {
+    const parameters =
+      (values as any[]).length === 0
+        ? fns.map(() => [])
+        : (values as any[]).length === 1
+        ? fns.map(() => values)
+        : (values as any[]).map((value) => array(value));
 
-      return isPromise<ReadonlyArray<V>>(waiter)
-        ? waiter.then((results) => concat(results, fn(...parameters)))
-        : concat(waiter, fn(...parameters));
-    }, [] as any);
+    return fns.reduce(
+      (waiter, fn, index) =>
+        isPromise<ReadonlyArray<V>>(waiter)
+          ? waiter.then((results) => concat(results, fn(...parameters[index])))
+          : concat(waiter, fn(...parameters[index])),
+      [] as any,
+    );
+  };